requirements.txt
streamlit
pymupdf
sentence-transformers
faiss-cpu
requests
numpy
opencv-python
numpy
yt-dlp
openai-whisper
pytube
whisper
